+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Can use `ImprovedManager(allow_bulk=False)` to disable bulk operations
 - [dj] Added `generate_json_streaming_content` to `ImprovedViewSet` to stream JSON response
   - Added `parse_streaming_response` in test utils to handle streaming responses in unittests
+  - Updated `APITestCase.build_url` to use the endpoint's name like `reverse` and also handle query params
 
 ---